# nodejs-server-server/api/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Airline Management App
host: airlinemanagement.azurewebsites.net
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /help:
    x-swagger-router-controller: help
    get:
      operationId: getHelpPage
      description: Get Help Page
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /about:
    x-swagger-router-controller: about
    get:
      operationId: getAboutPage
      description: Get About Page
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /passengers:
    x-swagger-router-controller: passenger
    get:
      operationId: getAllPassengers
      description: Get all passengers
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPassengersListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: newPassenger
      description: Add a new passenger
      parameters:
        - name: body
          description: Passenger properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Passenger"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /passengers/{id}:
    x-swagger-router-controller: passenger
    get:
      operationId: getPassengerById
      description: Get a passenger by id
      parameters:
        - name: id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetPassengerResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      operationId: updatePassenger
      description: Update passenger info
      parameters:
        - name: id
          description: Passenger id
          type: number
          in: path
          required: true
        - name: body
          description: Passenger properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Passenger"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      operationId: patchPassenger
      description: Assign a passenger to a flight and update passenger flight_no and assigned fields
      parameters:
        - name: id
          description: Passenger id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: delPassenger
      description: Delete a passenger by id
      parameters:
        - name: id
          description: Passenger id
          type: number
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /flights:
    x-swagger-router-controller: flight
    get:
      operationId: getAllFlights
      description: Get all Flights
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GetFlightListResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
        operationId: newFlight
        description: Add a new flight
        parameters:
          - name: body
            description: Flight properties
            in: body
            required: true
            schema:
              $ref: "#/definitions/Flight"
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GeneralResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  GetPassengersListResponse:
    required:
      - passengers
    properties:
      passengers:
        type: array
        items: 
          type: object
          properties:
            passenger_id:
              type: number
            first_name:
              type: string
            last_name:
              type: string
            destination:
              type: string
            flight_no:
              type: number
            assigned:
              type: boolean
  GetFlightListResponse:
    required:
      - flights
    properties:
      flights:
        type: array
        items: 
          type: object
          properties:
            flight_no:
              type: number
            airline_name:
              type: string
            departure_time:
              type: string
            departure_date:
              type: string
            destination:
              type: string
            no_passengers:
              type: number
            total_seats:
              type: number
  Passenger:
    type: object
    properties:
      first_name:
        type: string
        description: first name of the passenger
      last_name:
        type: string
        description: last name of the pssenger
      destination:
        type: string
        description: planned destination for the passenger
    required:
      - first_name
      - last_name
      - destination
  Flight:
    type: object
    properties:
      airline_name:
        type: string
        description: airline name for the flight
      departure_date:
        type: string
        description: date of the flight (mm-dd-yyyy)
      departure_time:
        type: string
        description: time of the flight (hh:MM TT)
      destination:
        type: string
        description: planned destination for the flight
      total_seats:
        type: number
        description: planned destination for the flight
    required:
      - airline_name
      - departure_date
      - departure_time
      - destination
      - total_seats
  GeneralResponse:
    type: object
    properties:
      success:
        type: number
        description: returns 1 if successful
      description:
        type: string
        description: general response 
    required:
      - success
      - description
  GetPassengerResponse:
    required:
      - passenger_id
      - first_name
      - last_name
      - destination
    properties:
      passenger_id:
        type: number
      first_name: 
        type: string
      last_name:
        type: string
      destination:
        type: string
      flight_no:
        type: number
      assigned:
        type: boolean
...